Feature/jssnmshd 2686 resolve todos (#6)
* feat: add notification selection for cordova

* feat: add notification selection to web

* chore: bump versions

* chore: comment

* fix: rename cordova file system interface

* feat: check rimraf call

* chore: update package-lock

* chore: comment

* chore: bump version

* feat: catch notification error on web

* fix: sanity check rimraf

* chore: bump versions

* chore: bump common version

* chore: update package-locks

* fix: error handling

+++ packages/electron/src/ElectronFileAccess.ts
@@ -205,7 +205,16 @@ export class ElectronFileAccess implements INativeFileAccess {
     public async deleteDirectory(path: string, storage: NativeFileStorage = NativeFileStorage.Data): Promise<Result<void>> {
         try {
             const directoryPath = this.getStoragePath(path, storage);
-            // TODO: JSSNMSHDD-2686 (Check this rimraf call!)
+
+            if (
+                !directoryPath.startsWith(__dirname) &&
+                !directoryPath.startsWith(this.datadir) &&
+                !directoryPath.startsWith(this.homedir) &&
+                !directoryPath.startsWith(this.tmpdir)
+            ) {
+                return Result.fail(this.resolveErrorCode("EACCES"));
+            }
+
             await new Promise((resolve) => rimraf(directoryPath, resolve));
             return Result.ok(undefined);
         } catch (err) {
